# Optimize computeElemValues for constant monomials