--- a/framework/include/base/MooseVariable.h
+++ b/framework/include/base/MooseVariable.h
@@ -345,15 +345,15 @@ public:
   /**
    * Compute values at facial quadrature points
    */
-  void computeElemValuesFace();
+  virtual void computeElemValuesFace();
   /**
    * Compute values at facial quadrature points for the neighbor
    */
-  void computeNeighborValuesFace();
+  virtual void computeNeighborValuesFace();
   /**
    * Compute values at quadrature points for the neighbor
    */
-  void computeNeighborValues();
+  virtual void computeNeighborValues();
   /**
    * Compute nodal values of this variable
    */