--- a/framework/src/base/SystemBase.C
+++ b/framework/src/base/SystemBase.C
@@ -507,11 +507,13 @@ SystemBase::addVariable(const std::string & var_name,
   {
     // FIXME: we cannot refer fetype in libMesh at this point, so we will just make a copy in
     // MooseVariableBase.
-    auto * var =
-        (type == (FEType(0, MONOMIAL)))
-            ? new MooseVariableConstMonomial(
-                  var_num, type, *this, _subproblem.assembly(tid), _var_kind)
-            : new MooseVariable(var_num, type, *this, _subproblem.assembly(tid), _var_kind);
+    MooseVariableBase * var;
+    if (type == FEType(0, MONOMIAL))
+      var = new MooseVariableConstMonomial(
+          var_num, type, *this, _subproblem.assembly(tid), _var_kind);
+    else
+      var = new MooseVariable(var_num, type, *this, _subproblem.assembly(tid), _var_kind);
+
     var->scalingFactor(scale_factor);
     _vars[tid].add(var_name, var);
   }