- Create `MooseVariableConstMonomial` that overrides `computeElemValues` and makes use of fact that there's only one DOF per element and the elemental value is the same at every quadrature point

Performance summary (with iprofiler) for following input file:

- Old code:
    - Total run time: 7.745 s
    - Time in `MooseVariable::computeElemValues`: 1.256 s
    - % of calculation: 16

- New code:
    - Total run time: 5.766 s
    - Time in `MooseVariableConstMonomial::computeElemValues`: 0.624 s
    - % of calculation : 11

So looks about like a 2x speed-up.

```
[Mesh]
  type = GeneratedMesh
  dim = 2
  nx = 10
  ny = 10
[]

[Variables]
  [./u]
    family = MONOMIAL
    order = CONSTANT
  [../]
[]

[Kernels]
  [./rxn]
    type = Reaction
    variable = u
  [../]
  [./ffn]
    type = UserForcingFunction
    variable = u
    function = 1
  [../]
[]

[Executioner]
  type = Steady

  # Preconditioned JFNK (default)
  solve_type = 'JFNK'
[]

[LotsOfAuxVariables]
  [./avar]
    number = 20000
    family = MONOMIAL
    order = CONSTANT
  [../]
[]
```

Targets #9836