@lindsayad 2017-09-13
There's still some clean-up I can do in `computeElemValues`...