@andrsd 2017-09-13
Also benchmark a 3D problem: more q-points per element, you should see a bigger gain.