@lindsayad 2017-09-14
Ready for further review.