@lindsayad 2017-09-14
Yea, it should be closed.